% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotateFeature.R
\name{rotateFeature}
\alias{rotateFeature}
\title{Rotate A Geographic Feature}
\usage{
rotateFeature(Polygon, Age = 0)
}
\arguments{
\item{Polygon}{An rgdal polygon}

\item{Age}{A numeric value; A whole number in millions of years}
}
\value{
An rgdal spatial object
}
\description{
Rotate a geographic rgdal polygon to its paleocoordinates
}
\details{
Rotate a rgdal spatial object to its paleogeographic coordinates under an EarthByte rotation model using a custom data service powered by the Macrostrat and GPlates development teams. This is the same rotation model returned by default through the Paleobiology Database. The data service is, in principle, capable of rotating individual points or multiple features. However, this wrapper function will only accept a single feature.

The rotation model used is Wright, N, S. Zahirovic, RD Müller, and M Seton (2013) "Towards community-driven, open-access paleogeographic reconstructions: integrating open-access paleogeographic and paleobiology data with plate tectonics" \emph{Biogeosciences} 10:1529-1541.
}
\examples{

# Download a polygon of Dane County, Wisconsin, United States, North America
# DaneCounty<-downloadWOF(Place="Dane",Type="county")

# Download a polygon of Wisconsin, United States, North America
# Wisconsin<-downloadWOF(Place="Wisconsin",Type="region")

# Rotate Dane county to its position during the Danian stage (65 million years ago)
# DaneDanian<-rotateFeature(Polygon=DaneCounty,Age=65)

}
\author{
Andrew A. Zaffos
}

